Move models folder out of utils
