+++ commands/mod/mentionaccess.js
@@ -1,5 +1,5 @@
 const commando = require('discord.js-commando');
-const mentionaccess = require('../../utils/models/mentionaccess.js');
+const mentionaccess = require('../../models/mentionaccess.js');
 
 module.exports = class MentionAccessCommand extends commando.Command {
 	constructor(client) {
